+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Environment variable substitution in INI config files with `${VAR}` and `${VAR:-default}` syntax (closes [#362](https://github.com/netresearch/ofelia/issues/362))
+
 ## [0.21.5] - 2026-03-18
 
 ### Added